refactor(protocol): split communication contracts
